refactor: App.tsx 리팩토링
- app.tsx 코드를 한눈에 보기 위해  util화 및 template 폴더로 함수, 컴포넌트 분리
