+++ client/src/common/utils/index.ts
@@ -1,5 +1,6 @@
 import * as api from './api';
 import * as uriParser from './uriParser';
 import { blockPage } from './blockPage';
+import registerToken from './registerToken';
 
-export { api, uriParser, blockPage };
+export { api, uriParser, blockPage, registerToken };
