+++ client/src/shared/App.tsx
@@ -1,33 +1,12 @@
 import React, { Fragment, useEffect } from 'react';
-import styled from 'styled-components';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import { Chatroom, Login } from '@pages/index';
 import { Header, Sidebar } from '@components/organisms';
-import { api, uriParser, blockPage } from '@utils/index';
+import { registerToken, blockPage } from '@utils/index';
 import { getUserInfo, getUserChatroom } from '@dispatch/index';
-
-const Main = styled.div`
-  display: flex;
-  width: 100%;
-  height: 94%;
-`;
-const MainBox = styled.div`
-  width: 77%;
-  height: 100%;
-`;
+import { Main, MainBox } from '@components/templates';
 
 const App = () => {
-  const registerToken = async () => {
-    if (uriParser.isExistParseCode()) {
-      const code = uriParser.getCode();
-      const token = await api.getToken(code);
-      if (token) {
-        localStorage.setItem('token', token);
-        window.location.href = '/';
-      }
-    }
-  };
-
   useEffect(() => {
     registerToken().then(() => {
       blockPage();
